optimized check_hit and nearby_objects functions
Now the distances between the robots and robots and shots within their radius of view (which is minimum 2 times the robot radius) is computed for all the robots at a single call. The distances with the nearby objects is also outputted in the 2 dictionaries "nearby_robots" and "nearby_shots" and is used from the check_hit function that now does not need to recompute all the distances. This speeds up the rendering making it smoother.